Quality of life improvements, elevator improved
- Changed from float to decimal types to increase accuracy for important value items.

- Made elevator move and collect value in an appropriate manner.

+++ IdleMiner/Assets/Scripts/StringFormatHelper.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public static class StringFormatHelper {
 
 	// Use this for initialization
-	public static string GetCurrencyString (float value) {
+	public static string GetCurrencyString (decimal value) {
 
         string customString = value.ToString("F2");
         return customString + " a";
